cambio de int en el id a Long

+++ src/main/java/com/example/hamacasbackend/entidades/hamacas/Hamaca.java
@@ -14,7 +14,7 @@ public class Hamaca {
 
     @Id // Designa este campo como la clave primaria
     @GeneratedValue(strategy = GenerationType.IDENTITY) // Configura la generación automática del valor
-    private int idHamaca;
+    private Long idHamaca;
 
     @Column // Es opcional si el nombre del campo coincide con el nombre de la columna en la DB
     private double precio;
