+++ src/main/java/com/example/hamacasbackend/entidades/reservas/Reserva.java
@@ -20,7 +20,7 @@ import java.time.LocalDateTime;
 public class Reserva {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idReserva;
+    private Long idReserva;
 
     @ManyToOne
     @JoinColumn(name = "idHamaca", referencedColumnName = "idHamaca")
